+++ src/main/java/com/example/QuizSystemProject/dto/UserResponse.java
@@ -6,7 +6,7 @@ import com.example.QuizSystemProject.Model.User; // Entity'den dönüşüm için
 // Hassas bilgiler (parola gibi) dahil EDİLMEZ.
 public class UserResponse {
 
-    private Long id;
+    private int id;
     private String username;
     private String name;
     private String surname;
@@ -38,8 +38,8 @@ public class UserResponse {
     // Getter ve Setterlar (Setterlar API'den veri almadığı için zorunlu değildir ama eklenebilir)
     // IDE ile otomatik oluşturabilirsiniz.
 
-    public Long getId() { return id; }
-    public void setId(Long id) { this.id = id; } // Yanıt DTO'larında setterlar genellikle kullanılmaz ama bulunabilir
+    public int getId() { return id; }
+    public void setId(int id) { this.id = id; } // Yanıt DTO'larında setterlar genellikle kullanılmaz ama bulunabilir
 
     public String getUsername() { return username; }
     public void setUsername(String username) { this.username = username; }
